refactor: web 모듈로 컨트롤러 구성요소 이동
- Controller, AcceptanceTest 이동
