+++ web/build.gradle
@@ -5,6 +5,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.flywaydb:flyway-core'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
     testImplementation 'io.rest-assured:rest-assured:3.3.0'
     runtimeOnly 'com.h2database:h2'
 }
