+++ web/src/test/java/kitchenpos/ApplicationTest.java
