+++ web/src/test/java/kitchenpos/menugroup/ui/MenuGroupAcceptanceTestSupport.java
@@ -1,4 +1,4 @@
-package kitchenpos.menugroup;
+package kitchenpos.menugroup.ui;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
