+++ web/src/test/java/kitchenpos/order/ui/OrderAcceptanceTestSupport.java
@@ -1,4 +1,4 @@
-package kitchenpos.order;
+package kitchenpos.order.ui;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
