+++ web/src/test/java/kitchenpos/product/ui/ProductAcceptanceTestSupport.java
@@ -1,4 +1,4 @@
-package kitchenpos.product;
+package kitchenpos.product.ui;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
