+++ web/src/test/java/kitchenpos/tablegroup/ui/TableGroupAcceptanceTestSupport.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup;
+package kitchenpos.tablegroup.ui;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
